requirements.txt
streamlit==1.31.1
pymysql==1.1.0
pandas==2.2.0
python-dotenv==1.0.1
boto3==1.34.39
sqlalchemy==2.0.23